# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This contains the commn REST api to support a modern cricket scoreboard project.",
        "title": "Cricket Scoreboard API",
        "termsOfService": "https://github.com/dashu-baba/cricket-scoreboard-api",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/dashu-baba/cricket-scoreboard-api/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/series": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Create a series item",
                "parameters": [
                    {
                        "description": "Create Series",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.SeriesCreateModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Series"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
        "/series/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Get singe item of series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responsemodels.Team"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Update series status",
                "parameters": [
                    {
                        "description": "Update Series Status Model",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.UpdateSeriesStatusModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
        "/series/:id/matches": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Create list of matches under a series",
                "parameters": [
                    {
                        "description": "Create Matches",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.MatchCreateModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
        "/series/:id/matches/:matchid": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Update starting players of the match",
                "parameters": [
                    {
                        "description": "Match Playing Squad",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.MatchPlayingSquadModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Update match status",
                "parameters": [
                    {
                        "description": "Update Match Status Model",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.UpdateSeriesStatusModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
        "/series/:id/matches/:matchid/innings": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Created an innings",
                "parameters": [
                    {
                        "description": "Create Innings Model",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.CreateInningsModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
        "/series/:id/teams": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Add list of teams to the series item",
                "parameters": [
                    {
                        "description": "Add Teams",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.TeamsAddModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Series"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Remove list of teams to the series item",
                "parameters": [
                    {
                        "description": "Remove Teams",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.TeamsRemoveModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
        "/series/:id/teams/:teamid": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "description": "Update Squad model",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.UpdateSquadModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        },
        "requestmodels.CreateInningsModel": {
            "type": "object",
            "required": [
                "battingTeamId",
                "bowlingTeamId",
                "tossWinningTeamId"
            ],
            "properties": {
                "battingTeamId": {
                    "type": "string"
                },
                "bowlingTeamId": {
                    "type": "string"
                },
                "tossWinningTeamId": {
                    "type": "string"
                }
            }
        },
...
        "requestmodels.MatchCreateModel": {
            "type": "object",
            "required": [
                "matches"
            ],
            "properties": {
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requestmodels.Match"
                    }
                }
            }
        },
        "requestmodels.MatchPlayingSquadModel": {
            "type": "object",
            "required": [
                "teamId"
            ],
            "properties": {
                "extras": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "players": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "teamId": {
                    "type": "string"
                }
            }
        },
...
        "requestmodels.SeriesCreateModel": {
            "type": "object",
            "required": [
                "gameType",
                "name",
                "teams"
            ],
            "properties": {
                "gameType": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requestmodels.SeriesParticipantModel"
                    }
                }
            }
        },
...
        "requestmodels.TeamsAddModel": {
            "type": "object",
            "required": [
                "teams"
            ],
            "properties": {
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requestmodels.SeriesParticipantModel"
                    }
                }
            }
        },
        "requestmodels.TeamsRemoveModel": {
            "type": "object",
            "required": [
                "teams"
            ],
            "properties": {
                "teams": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "requestmodels.UpdateSeriesStatusModel": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "integer"
                }
            }
        },
        "requestmodels.UpdateSquadModel": {
            "type": "object",
            "properties": {
                "addedPlayer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "removedPlayer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "responsemodels.ErrorModel": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responsemodels.Series": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodels.Team"
                    }
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "responsemodels.Team": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodels.Player"
                    }
                }
            }
        }
    }
}